Marked several intentionally unreferenced variables. Added CoordSystemType enumeration. Doweling on Direct3D camera test. Fixed cast warnings. Fixed D3D verify macro bug.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@763 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ test/testasyncsocket.cpp
@@ -77,7 +77,7 @@ DWORD Sender::Run(void *userData) {
 
         AssertNoException("Socket::Cleanup", Socket::Cleanup());
 
-        return cnt;
+        return static_cast<DWORD>(cnt);
     } catch (...) {
         return 0;
     }
@@ -145,7 +145,7 @@ DWORD Receiver::Run(void *userData) {
 
         AssertNoException("Socket::Cleanup", Socket::Cleanup());
 
-        return cnt;
+        return static_cast<DWORD>(cnt);
     } catch (...) {
         return 0;
     }
